pyproject.toml
[tool.poetry]
name = "afd"
version = "0.1.0"
description = ""
authors = ["Oliver Ni <oliver.ni@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.1.2"
starlette = "^0.14.2"
uvicorn = "^0.13.4"
httpx = "^0.17.1"
Jinja2 = "^2.11.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100